+++ src/main/java/net/gamingeinstein/paxelsfordummies/registries/ModItems.java
@@ -4,6 +4,7 @@ import net.gamingeinstein.paxelsfordummies.PaxelsForDummies;
 import net.gamingeinstein.paxelsfordummies.item.PaxelItem;
 import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.Item.Properties;
 import net.minecraft.world.item.Tiers;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.neoforge.event.BuildCreativeModeTabContentsEvent;
@@ -15,23 +16,17 @@ public class ModItems {
 
     // Paxels
     public static final DeferredItem<PaxelItem> WOODEN_PAXEL = ITEMS.register("wooden_paxel",
-            () -> new PaxelItem(Tiers.WOOD, new Item.Properties()
-                    .attributes(PaxelItem.createAttributes(Tiers.WOOD, 6.0f, -3.2f))));
+            () -> new PaxelItem(Tiers.WOOD, 6.0f, -3.2f, new Item.Properties()));
     public static final DeferredItem<PaxelItem> STONE_PAXEL = ITEMS.register("stone_paxel",
-            () -> new PaxelItem(Tiers.STONE, new Item.Properties()
-                    .attributes(PaxelItem.createAttributes(Tiers.STONE, 7.0f, -3.2f))));
+            () -> new PaxelItem(Tiers.STONE, 7.0f, -3.2f, new Properties()));
     public static final DeferredItem<PaxelItem> IRON_PAXEL = ITEMS.register("iron_paxel",
-            () -> new PaxelItem(Tiers.IRON, new Item.Properties()
-                    .attributes(PaxelItem.createAttributes(Tiers.IRON, 6.0f, -3.0f))));
+            () -> new PaxelItem(Tiers.IRON, 6.0f, -3.0f, new Item.Properties()));
     public static final DeferredItem<PaxelItem> GOLDEN_PAXEL = ITEMS.register("golden_paxel",
-            () -> new PaxelItem(Tiers.GOLD, new Item.Properties()
-                    .attributes(PaxelItem.createAttributes(Tiers.GOLD, 6.0f, -3.1f))));
+            () -> new PaxelItem(Tiers.GOLD, 6.0f, -3.1f, new Item.Properties()));
     public static final DeferredItem<PaxelItem> DIAMOND_PAXEL = ITEMS.register("diamond_paxel",
-            () -> new PaxelItem(Tiers.DIAMOND, new Item.Properties()
-                    .attributes(PaxelItem.createAttributes(Tiers.DIAMOND, 5.0f, -3.0f))));
+            () -> new PaxelItem(Tiers.DIAMOND, 5.0f, -3.0f, new Item.Properties()));
     public static final DeferredItem<PaxelItem> NETHERITE_PAXEL = ITEMS.register("netherite_paxel",
-            () -> new PaxelItem(Tiers.NETHERITE, new Item.Properties().fireResistant()
-                    .attributes(PaxelItem.createAttributes(Tiers.NETHERITE, 5.0f, -3.0f))));
+            () -> new PaxelItem(Tiers.NETHERITE, 5.0f, -3.0f, new Item.Properties().fireResistant()));
 
     public static void register(IEventBus eventBus) {
         ITEMS.register(eventBus);