+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
 
+## Version 0.1.1
+
+### Bug Fixes
+
+* Fixed Durability not being correctly set on Paxels
+---
+
 ## Version 0.1.0
 
 Initial release
 
 ### Additions
 
-- Paxels for each Vanilla material (Wood, Stone, Iron, Gold, Diamond, and Netherite)
+* Paxels for each Vanilla material (Wood, Stone, Iron, Gold, Diamond, and Netherite)